% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{calc_mean_ar}
\alias{calc_mean_ar}
\title{Calculate mean allelic richness}
\usage{
calc_mean_ar(genind)
}
\arguments{
\item{genind}{genind}
}
\value{
allelic richness averaged across all loci
}
\description{
Calculate mean allelic richness
}
