% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{calc_mean_biar}
\alias{calc_mean_biar}
\title{Calculate mean allelic richness for biallelic data}
\usage{
calc_mean_biar(dos)
}
\arguments{
\item{dos}{dosage matrix}
}
\value{
allelic richness averaged across all loci
}
\description{
Calculate mean allelic richness for biallelic data
}
