% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{calc_mean_het}
\alias{calc_mean_het}
\title{Calculate mean heterozygosity}
\usage{
calc_mean_het(hetmat)
}
\arguments{
\item{hetmat}{matrix of heterozygosity (0 = homozygote, 1 = heterozygote)}
}
\value{
heterozygosity averaged across all individuals and then all loci
}
\description{
Calculate mean heterozygosity
}
\keyword{internal}
