% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{calc_pi}
\alias{calc_pi}
\title{Calculate nucleotide diversity (pi) from dosage data}
\usage{
calc_pi(dos, nloci = NULL)
}
\arguments{
\item{dos}{a ni X nl dosage matrix containing the number of derived/alternate alleles each individual carries at each SNP}

\item{L}{length of the sequence (\emph{note:} defaults to number of loci in the provided dosage matrix; TODO: COME BACK AND FIX THIS)}
}
\description{
Wrapper for \link[hierfstat]{pi.dosage} function
}
\keyword{internal}
