% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{check_data}
\alias{check_data}
\title{Check coordinate and genetic data}
\usage{
check_data(gen, coords)
}
\arguments{
\item{gen}{genetic data}

\item{coords}{coordinates}
}
\description{
Check that the number of individuals in each data set align
}
\keyword{internal}
