% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{coord_proj}
\alias{coord_proj}
\title{TEMP COORDS FUNCTION}
\usage{
coord_proj(coords, spdf, crop_to_spdf = FALSE)
}
\arguments{
\item{coords}{coordinates}

\item{spdf}{SpatialPointsDataFrame}

\item{crop_to_spdf}{whether to crop the coordinates to the spdf}
}
\description{
TEMP COORDS FUNCTION
}
