% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_raster.R
\name{coords_to_raster}
\alias{coords_to_raster}
\title{Create RasterLayer from coordinates}
\usage{
coords_to_raster(coords, buffer = 0, agg = NULL, disagg = NULL, plot = FALSE)
}
\arguments{
\item{coords}{coordinates (two columns, the first should be x and the second should be y)}

\item{buffer}{buffer to add to edge of raster (defaults to 0)}

\item{agg}{aggregation factor to apply to raster (defaults to NULL)}

\item{disagg}{disaggregation factor to apply to raster (defaults to NULL)}

\item{plot}{whether to plot resulting raster with coords (defaults to FALSE)}
}
\value{
RasterLayer
}
\description{
Create RasterLayer from coordinates
}
\examples{
load_mini_ex()
coords_to_raster(mini_coords, buffer = 10, plot = TRUE)
}
