% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gd.R
\name{dist_general}
\alias{dist_general}
\title{General function for dist_gd
TODO: CREATE GENERAL CIRCLE GD AND RESIST GD}
\usage{
dist_general(
  x,
  coords,
  lyr,
  stat,
  maxdist,
  distmat,
  rarify = FALSE,
  rarify_n = 2,
  rarify_nit = 5,
  min_n = 2,
  fun = mean,
  L = "nvariants",
  rarify_alleles = TRUE,
  parallel = FALSE,
  ncores = NULL,
  crop_edges = FALSE,
  ...
)
}
\arguments{
\item{x}{data to be summarized by the moving window (\emph{note:} order matters! \code{coords} should be in the same order, there are currently no checks for this). The class of \code{x} required depends on the statistic being calculated (see the \code{stat} argument and the function description for more details)}

\item{stat}{moving window statistic to calculate (can either be \code{pi} for nucleotide diversity (\code{x} must be a dosage matrix), \code{Ho} for average observed heterozygosity across all loci (\code{x} must be a heterozygosity matrix) , "allelic_richness" for average allelic richness across all loci (\code{x} must be a \code{genind} type object), "biallelic_richness" to get average allelic richness across all loci for a biallelic dataset (\code{x} must be a dosage matrix). \code{stat} can also be set to any function that will take \code{x}as input and return a single numeric value (for example, \code{x} can be a vector and \code{stat} can be set equal to a summary statistic like \code{mean}, \code{sum}, or \code{sd})}

\item{...}{if a function is provided for \code{stat}, additional arguments to pass to the \code{stat} function (e.g. if \code{stat = mean}, users may want to set \code{na.rm = TRUE})}
}
\value{
SpatRaster that includes a raster layer of genetic diversity and a raster layer of the number of samples within the window for each cell
}
\description{
General function for dist_gd
TODO: CREATE GENERAL CIRCLE GD AND RESIST GD
}
