% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{div_mask}
\alias{div_mask}
\title{Mask diversity map based on sample number}
\usage{
div_mask(x, min_n, plot = FALSE, bkg.col = "white")
}
\arguments{
\item{x}{RasterStack where the first layer is genetic diversity and the second layer is sample count}

\item{min_n}{minimum number of samples (everything less than this number is masked)}

\item{plot}{whether to plot results}

\item{bkg.col}{background color for plotting map}
}
\value{
RasterLayer
}
\description{
Mask diversity map based on sample number
}
