% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_coords}
\alias{ex_coords}
\title{Example coordinates}
\format{
A data frame with 100 rows and 2 columns
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\usage{
ex_coords
}
\description{
A data set of coordinates
}
\keyword{datasets}
