% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_gd.R
\name{geo_dist}
\alias{geo_dist}
\title{Calculate geographic distance between coordinates}
\usage{
geo_dist(coords, type = "Euclidean", lyr = NULL)
}
\arguments{
\item{coords}{dataframe with x and y coordinates}

\item{type}{The type of geographic distance to be calculated; options are "Euclidean" for direct distance, "topographic" for topographic distances, and "resistance" for resistance distances.}

\item{lyr}{DEM raster for calculating topographic distances or resistance raster for calculating resistance distances (RasterLayer or SpatRaster object)}
}
\value{
A distance matrix
}
\description{
Calculate geographic distance between coordinates
}
\details{
Euclidean, or linear, distances are calculated using the geodist package: Padgham M, Sumner M (2021). geodist: Fast, Dependency-Free Geodesic Distance Calculations. R package version 0.0.7, Available: https://CRAN.R-project.org/package=geodist.
Topographic distances are calculated using the topoDistance package: Wang I.J. (2020) Topographic path analysis for modeling dispersal and functional connectivity: calculating topographic distances using the TOPODISTANCE R package. Methods in Ecology and Evolution, 11: 265-272.
Resistance distances are calculated using the gdistance package: van Etten, J. (2017). R package gdistance: Distances and routes on geographical grids. Journal of Statistical Software, 76(1), 1–21.
}
