% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{get_adj}
\alias{get_adj}
\title{Helper function to get adjacent cells to a given cell index}
\usage{
get_adj(i, r, n, coord_cells)
}
\arguments{
\item{i}{cell index}

\item{r}{RasterLayer}

\item{n}{neighborhood matrix}

\item{coord_cells}{cell numbers of coordinates}
}
\value{
indices of coordinates that are adjacent to the given cell
}
\description{
Helper function to get adjacent cells to a given cell index
}
\keyword{internal}
