% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resist_gd.R
\name{get_resdist}
\alias{get_resdist}
\title{Get a matrix of resistance distances for \link[wingen]{resist_gd}}
\usage{
get_resdist(
  coords,
  lyr,
  fact = 0,
  ncores = 1,
  parallel = TRUE,
  progress = TRUE
)
}
\arguments{
\item{coords}{coordinates of samples as sf points, a two-column matrix, or a data.frame representing x and y coordinates (see Details for important information about projections)}

\item{lyr}{conductivity layer (higher values should mean greater conductivity) to move window across. Can be either a SpatRaster or RasterLayer.}

\item{fact}{aggregation factor to apply to \code{lyr} (defaults to 0; \emph{note:} increasing this value reduces computational time)}

\item{ncores}{if parallel = TRUE, number of cores to use for parallelization (defaults to total available number of cores minus 1)}

\item{parallel}{whether to parallelize the function (defaults to FALSE)}
}
\value{
a distance matrix used by \link[wingen]{resist_gd}
}
\description{
Create a distance matrix based on coordinates and a connectivity layer.
The output is a distance matrix where rows represent cells on the landscape
and columns represent individual locations on the landscape. Each value is
the resistance distance between each individual and each cell calculated
using the gdistance package. This matrix is used by \link[wingen]{resist_gd}.
}
\examples{
\dontrun{
load_mini_ex()
distmat <- get_resdist(mini_coords, mini_lyr)
}
}
