% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_gd.R
\name{ggplot_count}
\alias{ggplot_count}
\title{Plot moving window map of sample counts}
\usage{
ggplot_count(x, index = NULL, breaks = 100, col = viridis::mako(breaks), ...)
}
\arguments{
\item{x}{single SpatRaster of counts or SpatRaster where indexed layer is sample counts}

\item{index}{if a raster stack is provided, index of the sample count layer to plot
(assumes this is a stacked output from window_gd and defaults to plotting the last layer (which should be sample counts))}

\item{col}{color palette to use for plotting (defaults to viridis::magma palette)}
}
\value{
list of ggplots
}
\description{
Plot sample counts layer produced by \link[wingen]{window_gd} or \link[wingen]{krig_gd}
}
\examples{
data("mini_lyr")
plot_count(mini_lyr)
}
