% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_gd.R
\name{ggplot_gd}
\alias{ggplot_gd}
\title{Plot moving window map of genetic diversity}
\usage{
ggplot_gd(x, bkg = NULL, index = NULL, col = viridis::magma(100))
}
\arguments{
\item{x}{output from \link[wingen]{window_gd} or \link[wingen]{krig_gd} (RasterStack where first layer is genetic diversity)}

\item{bkg}{optional raster or sf polygon}

\item{index}{index of raster layers to plot (defaults to plotting all of the layers except the one called "sample_count", if more than one layer is provided)}

\item{col}{color palette to use for plotting (defaults to \link[viridis]{magma} palette)}
}
\value{
list of ggplots
}
\description{
Plot genetic diversity layer produced by \link[wingen]{window_gd} or \link[wingen]{krig_gd}
}
\examples{
data("mini_lyr")
ggplot_gd(mini_lyr)

}
