% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{helper_calc_ar}
\alias{helper_calc_ar}
\title{Helper function to calculate allelic richness}
\usage{
helper_calc_ar(genind)
}
\arguments{
\item{genind}{genind object}
}
\value{
allelic richness
}
\description{
Helper function to calculate allelic richness
}
