% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{helper_calc_biar}
\alias{helper_calc_biar}
\title{Helper function to calculate allelic richness for a biallelic locus}
\usage{
helper_calc_biar(loc)
}
\arguments{
\item{loc}{genotypes at a biallelic locus (must have values of 0, 1, or 2)}
}
\description{
Helper function to calculate allelic richness for a biallelic locus
}
\keyword{internal}
