% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{krig_div}
\alias{krig_div}
\title{Krige map}
\usage{
krig_div(r, spdf, xy = TRUE, n_cell = 10000)
}
\arguments{
\item{r}{raster for kriging}

\item{spdf}{SpatialPointsDataFrame object to create grid for kriging}

\item{xy}{whether to co-krige with x and y (~x+y)}

\item{n_cell}{number of cells to interpolate across}
}
\value{
Raster with kriged values
}
\description{
Krige map
}
