% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_gd.R
\name{krig_gd_lyr}
\alias{krig_gd_lyr}
\title{Krige RasterLayer}
\usage{
krig_gd_lyr(
  r,
  grd = NULL,
  coords = NULL,
  xy = FALSE,
  resample = FALSE,
  agg_grd = NULL,
  disagg_grd = NULL,
  agg_r = NULL,
  disagg_r = NULL,
  resample_first = TRUE,
  n_cell = 1000
)
}
\arguments{
\item{r}{raster for kriging}

\item{grd}{object to create grid for kriging, can be RasterLayer, SpatialPointsDataFrame, or a gridded object as defined by 'sp'. If undefined, will use \code{r} to create a grid.}

\item{coords}{if provided, kriging will occur based only on values at these coordinates}

\item{xy}{whether to co-krige with x and y (~x+y)}

\item{resample}{whether to resample grd or r. Set to "r" to resample r to grd Set to "grd" to resample grd to r. Defaults to FALSE but we \emph{highly recommend setting this to either "grd" or "r"}}

\item{agg_grd}{factor to use for aggregation of grd, if provided}

\item{disagg_grd}{factor to use for disaggregation of grd, if provided}

\item{agg_r}{factor to use for aggregation of r, if provided}

\item{disagg_r}{factor to use for disaggregation, of r if provided}

\item{resample_first}{if aggregation or disaggregation is used in addition to resampling, whether to resample before (resample_first = TRUE) or after (resample_first = FALSE) aggregation/disaggregation (defaults to TRUE)}

\item{n_cell}{number of cells to interpolate across if SpatialPointsDataFrame is provided for \code{grd}}
}
\value{
RasterLayer
}
\description{
Helper function for \code{\link{krig_gd}}
}
\keyword{internal}
