% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{krig_rast_lyr}
\alias{krig_rast_lyr}
\title{Krige RasterLayer}
\usage{
krig_rast_lyr(
  r,
  grd = NULL,
  xy = FALSE,
  agg = NULL,
  disagg = NULL,
  n_cell = 10000
)
}
\arguments{
\item{r}{raster for kriging}

\item{grd}{object to create grid for kriging, can be RasterLayer, SpatialPointsDataFrame, or a gridded object as defined by 'sp'. If undefined, will use \code{r} to create a grid.}

\item{xy}{whether to co-krige with x and y (~x+y)}

\item{agg}{factor used for aggregation if provided}

\item{disagg}{factor used for disaggregation if provided}

\item{n_cell}{number of cells to interpolate across if SpatialPointsDataFrame is provided for \code{grd}}
}
\value{
RasterLayer
}
\description{
Helper function for \code{\link{krig_rast}}
}
\keyword{internal}
