% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_earth.R
\name{load_mini_ex}
\alias{load_mini_ex}
\title{Load mini middle earth example}
\usage{
load_mini_ex()
}
\value{
three objects are assigned in the GlobalEnv (vcf, coords, and lyr)
}
\description{
Loads mini middle earth example data and assigns to simple names
}
