% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_distmat}
\alias{lotr_distmat}
\title{Middle earth example resistance distance matrix}
\format{
matrix
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
lotr_distmat
}
\description{
Resistance distance matrix made based on \link[wingen]{lotr_lyr} aggregated by a factor of 5 and \link[wingen]{lotr_coords}
}
\keyword{datasets}
