% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_lyr}
\alias{lotr_lyr}
\title{Middle earth example raster}
\format{
A RasterLayer of middle earth
}
\usage{
lotr_lyr
}
\description{
RasterLayer of middle earth from CITE
}
\keyword{datasets}
