% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_range}
\alias{lotr_range}
\title{Middle earth example range polygon}
\format{
SpatialPolygonsDataFrame
}
\usage{
lotr_range
}
\description{
SpatialPolygonsDataFrame of range map
}
\keyword{datasets}
