% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_vcf}
\alias{lotr_vcf}
\title{Middle earth example vcf}
\format{
Object of class vcfR with 100 individuals and 1000 loci
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
lotr_vcf
}
\description{
A Variant Call Format data set
}
\keyword{datasets}
