% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_gd.R
\name{mask_gd}
\alias{mask_gd}
\title{Mask genetic diversity map}
\usage{
mask_gd(x, mask, resample = "mask", minval = NULL, maxval = NULL)
}
\arguments{
\item{x}{Raster object to mask}

\item{mask}{Raster object or Spatial object to use as mask}

\item{resample}{if x and mask are non matching rasters, which layer to resample to match them (defaults to "y")}

\item{minval}{if mask is a Raster object, value of mask below which to mask}

\item{maxval}{if mask is a Raster object, value of mask above which to mask}
}
\value{
RasterLayer
}
\description{
Mask genetic diversity map
}
\examples{
library("raster")
load_mini_ex()
wpi <- window_gd(mini_vcf, mini_coords, mini_lyr, nloci = 10, rarify = TRUE)
kpi <- krig_gd(wpi, mini_lyr)
mpi <- mask_gd(kpi, mini_lyr, minval = 0.01)
plot_gd(mpi, main = "Kriged and Masked Pi")

}
