% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{middle_earth_coords}
\alias{middle_earth_coords}
\title{Middle earth example coordinates}
\format{
A data frame with 100 rows and 2 columns
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\usage{
middle_earth_coords
}
\description{
Middle earth example coordinates
}
\keyword{datasets}
