% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{middle_earth_vcf}
\alias{middle_earth_vcf}
\title{Middle earth example vcf}
\format{
Object of class vcfR with 100 individuals and 1000 loci
}
\usage{
middle_earth_vcf
}
\description{
A Variant Call Format data set
}
\keyword{datasets}
