% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_coords}
\alias{mini_coords}
\title{Mini middle earth example coordinates}
\format{
A data frame with 10 rows and 2 columns
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
mini_coords
}
\description{
Mini middle earth example coordinates
}
\keyword{datasets}
