% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_lyr}
\alias{mini_lyr}
\title{Mini middle earth example raster}
\format{
A RasterLayer of middle earth.
}
\source{
Created from simulations in Bishop et al. (2023) based on Rose, Robert A. (2020) GIS & Middle Earth Presentation & Data Set. William & Mary. \url{https://doi.org/10.21220/RKEZ-X707}
}
\usage{
mini_lyr
}
\description{
Small RasterLayer of middle earth based on an example digital elevation model of Tolkien's
Middle Earth produced by the Center for Geospatial Analysis at William & Mary
(Robert, 2020).
}
\keyword{datasets}
