% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_vcf}
\alias{mini_vcf}
\title{Mini middle earth example vcf}
\format{
Object of class vcfR with 10 individuals and 10 loci
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
mini_vcf
}
\description{
A Variant Call Format data set
}
\keyword{datasets}
