% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_vcf_NA}
\alias{mini_vcf_NA}
\title{Mini middle earth example vcf with NA values}
\format{
Object of class vcfR with 10 individuals and 10 loci
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
mini_vcf_NA
}
\description{
A Variant Call Format data set with NA values
}
\keyword{datasets}
