% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gd.R
\name{plot_count}
\alias{plot_count}
\title{Plot sample counts}
\usage{
plot_count(
  x,
  col = viridis::mako(100),
  zlim = NULL,
  main = NULL,
  legend = TRUE,
  legend.width = 1,
  axis.args = list(cex.axis = 1)
)
}
\arguments{
\item{x}{output from \link[wingen]{window_gd} or \link[wingen]{krig_gd} (RasterStack where second layer is sample counts)}

\item{col}{color pallete to use for plotting (defaults to viridis::magma pallete)}

\item{zlim}{limits of the color scale values}

\item{main}{character. Main plot title}

\item{breaks}{number of breaks to use in color scale (defaults to 10)}
}
\value{
plot of sample counts
}
\description{
Plot sample counts
}
