% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gd.R
\name{plot_gd}
\alias{plot_gd}
\title{Plot genetic diversity results}
\usage{
plot_gd(
  x,
  bkg = NULL,
  index = NULL,
  col = viridis::magma(breaks),
  breaks = 10,
  main = NULL,
  box = FALSE,
  ...
)
}
\arguments{
\item{x}{output from \link[wingen]{window_gd} or \link[wingen]{krig_gd} (RasterStack where first layer is genetic diversity)}

\item{bkg}{RasterLayer or other spatial object that will be plotted as the "background" in gray}

\item{index}{if RasterStack is provided, index of the sample count layer to plot (defaults to plotting first layer)}

\item{col}{color pallete to use for plotting (defaults to viridis::magma pallete)}

\item{breaks}{number of breaks to use in color scale (defaults to 10)}

\item{main}{character. Main plot title}

\item{...}{Graphical parameters. Any argument that can be passed to \code{\link[fields]{image.plot}} and to base \code{plot}, 
        such as axes=FALSE, main='title', ylab='latitude'}
}
\value{
plot of genetic diversity
}
\description{
Plot genetic diversity results
}
\examples{
data("mini_lyr")
plot_gd(mini_lyr)

}
