% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_gd.R, R/window_gd.R
\name{preview_gd}
\alias{preview_gd}
\title{Generate preview of moving window and sample counts}
\usage{
preview_gd(lyr, coords, wdim, fact = 0, sample_count = TRUE, min_n = 0)

preview_gd(lyr, coords, wdim, fact = 0, sample_count = TRUE, min_n = 0)
}
\arguments{
\item{lyr}{RasterLayer to slide window across}

\item{coords}{coordinates (two columns, the first should be x and the second should be y)}

\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created}

\item{fact}{aggregation factor to apply to the RasterLayer (\emph{note:} increasing this value reduces computational time)}

\item{sample_count}{whether to create plot of sample counts for each cell (defaults to TRUE)}

\item{min_n}{min number of samples to use in calculations (any focal cell with a window containing less than this number of samples will be assigned a value of NA)}
}
\value{
plots example window and sample counts (if sample_count = TRUE)

plots example window and sample counts (if sample_count = TRUE)
}
\description{
Generate preview of moving window and sample counts

Generate preview of moving window and sample counts
}
