% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{range_to_grid}
\alias{range_to_grid}
\title{Make grid for kriging}
\usage{
range_to_grid(xrange, yrange, len)
}
\arguments{
\item{xrange}{range of xvalues}

\item{yrange}{range of yvalues}

\item{len}{length provided to expand.grid}
}
\description{
Make grid for kriging
}
