% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_div.R
\name{rarify_het}
\alias{rarify_het}
\title{Rarefaction function}
\usage{
rarify_het(ar_df, sub, rarify_nit = 10, rarify_n = 4, fun = mean)
}
\arguments{
\item{ar_df}{dataframe of allelic richness (\emph{note:} cols should be individuals, rows should be loci)}

\item{sub}{subset of row indices}

\item{rarify_nit}{number of iterations (i.e. random samples of size rarify_n to draw)}

\item{rarify_n}{number of points to use for each rarefaction iteration}

\item{fun}{function to use to calculate the mean (defaults to base R mean)}
}
\value{
rarified mean allelic richness for a subsample
}
\description{
Rarefaction function
}
\keyword{internal}
