% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_gd.R
\name{raster_to_grid}
\alias{raster_to_grid}
\title{Conver a raster to a grid}
\usage{
raster_to_grid(x)
}
\arguments{
\item{x}{RasterLayer}
}
\value{
gridded SpatialPixelsDataFrame
}
\description{
Conver a raster to a grid
}
\keyword{internal}
