% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_gd.R
\name{raster_transform}
\alias{raster_transform}
\title{Transform raster}
\usage{
raster_transform(
  r,
  grd,
  resample = FALSE,
  agg_grd = NULL,
  disagg_grd = NULL,
  agg_r = NULL,
  disagg_r = NULL,
  resample_first = TRUE
)
}
\arguments{
\item{r}{RasterLayer or RasterStack}

\item{grd}{object to create grid for kriging, can be RasterLayer, SpatialPointsDataFrame, or a gridded object as defined by 'sp'. If undefined, will use \code{r} to create a grid.}

\item{resample}{whether to resample grd or r. Set to "r" to resample r to grd Set to "grd" to resample grd to r. Defaults to FALSE but we \emph{highly recommend setting this to either "grd" or "r"}}

\item{agg_grd}{factor to use for aggregation of grd, if provided}

\item{disagg_grd}{factor to use for disaggregation of grd, if provided}

\item{agg_r}{factor to use for aggregation of r, if provided}

\item{disagg_r}{factor to use for disaggregation, of r if provided}

\item{resample_first}{if aggregation or disaggregation is used in addition to resampling, whether to resample before (resample_first = TRUE) or after (resample_first = FALSE) aggregation/disaggregation (defaults to TRUE)}
}
\value{
stack of transformed rasters
}
\description{
Transform raster
}
