% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale values from zero to one}
\usage{
rescale(x, x.min = NULL, x.max = NULL, new.min = 0, new.max = 1)
}
\arguments{
\item{x}{vector of numbers}

\item{x.min}{current min}

\item{x.max}{current max}

\item{new.min}{new min}

\item{new.max}{new max}
}
\value{
values rescales to new min and max
}
\description{
Rescale values from zero to one
}
