% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{return_stat}
\alias{return_stat}
\title{Helper function to get genetic diversity functions}
\usage{
return_stat(x)
}
\arguments{
\item{x}{genetic diversity statistic}
}
\value{
function corresponding with desired statistic
}
\description{
Helper function to get genetic diversity functions
}
\keyword{internal}
