% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_ar.R
\name{sample_ar}
\alias{sample_ar}
\title{Calculate mean allelic richness of a sample}
\usage{
sample_ar(ar_df, sub, fun = mean)
}
\arguments{
\item{ar_df}{dataframe of allelic richness (\emph{note:} cols should be individuals, rows should be loci)}

\item{sub}{row indices of subsample}

\item{fun}{function to use to calculate the mean (defaults to base R mean)}
}
\value{
mean allelic richness of a subsample
}
\description{
Calculate mean allelic richness of a sample
}
\keyword{internal}
