% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_gd.R
\name{spdf_to_grid}
\alias{spdf_to_grid}
\title{Make grid from Spatial Points Data Frame}
\usage{
spdf_to_grid(spdf, n_cell = 1000)
}
\arguments{
\item{spdf}{Spatial Points Dataframe to create grid for kriging}

\item{n_cell}{number of grid cells to use when kriging}
}
\value{
gridded SpatialPixelsDataFrame
}
\description{
Make grid from Spatial Points Data Frame
}
\note{
code from: https://stackoverflow.com/questions/43436466/create-grid-in-r-for-kriging-in-gstat
}
\keyword{internal}
