% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vcf.R
\name{vcf_check}
\alias{vcf_check}
\title{Check if an object is a vcf or a path to a vcf}
\usage{
vcf_check(x)
}
\arguments{
\item{x}{vcfR object or path to vcf}
}
\value{
vcf object
}
\description{
Check if an object is a vcf or a path to a vcf
}
\keyword{internal}
