% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vcf.R
\name{vcf_to_dosage}
\alias{vcf_to_dosage}
\title{Convert vcf to dosage matrix}
\usage{
vcf_to_dosage(x)
}
\arguments{
\item{x}{can either be an object of class 'vcfR' or a path to a .vcf file}
}
\value{
returns dosage matrix
}
\description{
Convert vcf to dosage matrix
}
\examples{
data("mini_vcf")
vcf_to_dosage(mini_vcf)

}
\keyword{internal}
