% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vcf.R
\name{vcf_to_genind}
\alias{vcf_to_genind}
\title{Wrapper for vcfR2genind function that assigns pops}
\usage{
vcf_to_genind(x, pops = NULL)
}
\arguments{
\item{x}{can either be an object of class 'vcfR' or a path to a .vcf file}

\item{pops}{if NULL (default), and there are no pops detected from the vcf, each individual is assigned its own pop. If FALSE then genind$pop is left NULL. Alternatively, a vector of population assignments for each individual can be provided}
}
\value{
returns genind object
}
\description{
Wrapper for vcfR2genind function that assigns pops
}
\keyword{data("ex_vcf")}
\keyword{internal}
\keyword{vcf_to_genind(ex_vcf)}
