% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdim.R
\name{wdim_check}
\alias{wdim_check}
\title{Helper function to check that wdim object is correctly assigned}
\usage{
wdim_check(wdim)
}
\arguments{
\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created}
}
\value{
corrected wdim
}
\description{
Helper function to check that wdim object is correctly assigned
}
\examples{
wdim_check(c(4,4))
wdim_check(4)

}
