% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdim.R
\name{wdim_to_mat}
\alias{wdim_to_mat}
\title{Helper function to convert wdim object to neighbor matrix}
\usage{
wdim_to_mat(wdim)
}
\arguments{
\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created}
}
\value{
neighborhood matrix
}
\description{
Helper function to convert wdim object to neighbor matrix
}
\examples{
wdim_to_mat(3)
wdim_to_mat(c(3,5))

}
\keyword{internal}
