% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_ar.R
\name{window_ar}
\alias{window_ar}
\title{Sliding window allelic richness}
\usage{
window_ar(
  ar_df,
  coords,
  lyr,
  fact = 0,
  wdim = 10,
  rarify = TRUE,
  rarify_n = 4,
  rarify_nit = 10,
  min_n = 2,
  fast = FALSE,
  plot_its = FALSE,
  plot_its_steps = 1
)
}
\arguments{
\item{ar_df}{dataframe of allelic richness (\emph{note:} rows should be individuals, columns should be loci)}

\item{coords}{coordinates (two columns, the first should be x and the second should be y and the order should be the same as ar_df),}

\item{lyr}{raster layer to slide window across}

\item{fact}{factor of aggregation to apply to the raster layer (reduces computational time)}

\item{wdim}{dimensions (height x width) of window, if one value is provided a square window is created}

\item{rarify}{whether to use rarefaction in calculations}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{min_n}{min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)}

\item{fast}{if TRUE uses Rcpp mean to calculate mean, which is slightly faster than base R (\emph{note:} compromises on numerical accuracy)}

\item{plot_its}{whether to produce a plot every plot_its_steps iterations to show progress}

\item{plot_its_steps}{when to plot an iteration if plot_its is true}
}
\value{
RasterStack that includes a raster of allelic richness and a raster of the number of samples used in the calculation of allelic richness for each cell
}
\description{
Sliding window allelic richness
}
