% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{window_helper}
\alias{window_helper}
\title{Helper function for window calculations}
\usage{
window_helper(
  i,
  lyr,
  gen,
  coord_cells,
  nmat,
  stat,
  rarify,
  rarify_n,
  rarify_nit,
  min_n,
  fun
)
}
\arguments{
\item{i}{cell index}

\item{lyr}{RasterLayer to slide window across}

\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{coord_cells}{cell indices for each coordinate}

\item{nmat}{neighbor matrix}

\item{rarify}{if rarify = TRUE, rarefaction is performed}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{min_n}{min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}
}
\description{
Helper function for window calculations
}
\keyword{internal}
