% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingen-package.R
\docType{package}
\name{wingen-package}
\alias{wingen}
\alias{wingen-package}
\title{wingen: Sliding Window Maps of Genetic Diversity}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualize genetic diversity across space using a sliding window approach with options for rarefaction and interpolation.
}
\author{
\strong{Maintainer}: Anusha Bishop \email{anusha.bishop@berkeley.edu} (\href{https://orcid.org/0000-0003-1731-8683}{ORCID})

Authors:
\itemize{
  \item Ian Wang \email{ianwang@berkeley.edu} (\href{https://orcid.org/0000-0003-2554-9414}{ORCID})
}

}
\keyword{internal}
