% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_gd.R
\name{ggplot_count}
\alias{ggplot_count}
\title{Plot moving window map of sample counts}
\usage{
ggplot_count(x, index = NULL, col = viridis::mako(100))
}
\arguments{
\item{x}{single SpatRaster of counts or SpatRaster where indexed layer is sample counts}

\item{index}{index of raster layers to plot (defaults to plotting the one called "sample_count", if more than one layer is provided)}

\item{col}{color palette to use for plotting (defaults to viridis::mako palette)}
}
\value{
list of ggplots
}
\description{
Plot sample counts layer produced by \link[wingen]{window_gd} or \link[wingen]{krig_gd}
}
\examples{
data("mini_lyr")
ggplot_count(mini_lyr)
}
