% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_gd.R
\name{krig_gd}
\alias{krig_gd}
\title{Raster interpolation using 'autoKrige'}
\usage{
krig_gd(
  r,
  grd = NULL,
  coords = NULL,
  xy = FALSE,
  agg = NULL,
  disagg = NULL,
  n_cell = 10000
)
}
\arguments{
\item{r}{RasterLayer or RasterStack}

\item{grd}{object to create grid for kriging, can be RasterLayer, SpatialPointsDataFrame, or a gridded object as defined by 'sp'. If undefined, will use \code{r} to create a grid.}

\item{coords}{if provided, kriging will occur based only on values at these coordinates}

\item{xy}{whether to co-krige with x and y (~x+y)}

\item{agg}{factor used for aggregation if provided}

\item{disagg}{factor used for disaggregation if provided}

\item{n_cell}{number of cells to interpolate across if SpatialPointsDataFrame is provided for \code{grd}}
}
\value{
RasterLayer or RasterStack
}
\description{
Raster interpolation using 'autoKrige'
}
