% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_earth.R
\name{load_middle_earth_ex}
\alias{load_middle_earth_ex}
\title{Load middle earth example}
\usage{
load_middle_earth_ex()
}
\value{
three objects are loaded (lotr_vcf, lotr_coords, and lotr_lyrs)
}
\description{
Loads middle earth example data and assigns to simple names
}
