% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_earth.R
\name{load_middle_earth_ex}
\alias{load_middle_earth_ex}
\title{Middle earth example}
\usage{
load_middle_earth_ex(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to hide message (defaults to FALSE)}
}
\value{
three objects are loaded (lotr_vcf, lotr_coords, and lotr_lyrs)
}
\description{
Loads middle earth example data
}
\examples{
load_middle_earth_ex()
}
