% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_earth.R
\name{load_mini_ex}
\alias{load_mini_ex}
\title{Mini middle earth example}
\usage{
load_mini_ex(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to hide message (defaults to FALSE)}
}
\value{
three objects are assigned in the GlobalEnv (vcf, coords, and lyr)
}
\description{
Loads mini middle earth example data
}
\examples{
load_mini_ex()
}
