% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_coords}
\alias{lotr_coords}
\title{Middle earth example coordinates}
\format{
A data frame with 100 rows and 2 columns
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
}
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
lotr_coords
}
\description{
Middle earth example coordinates
}
\keyword{datasets}
