% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_lyr}
\alias{lotr_lyr}
\title{Middle earth example raster}
\format{
RasterLayer
}
\usage{
lotr_lyr
}
\description{
RasterLayer of middle earth from Rose, Robert A. (2020) GIS & Middle Earth Presentation & Data Set. William & Mary. https://doi.org/10.21220/RKEZ-X707
}
\keyword{datasets}
