% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_lyr}
\alias{lotr_lyr}
\title{Middle earth example raster}
\format{
RasterLayer
}
\source{
created from simulations in Bishop et al. (2023) based on Rose, Robert A. (2020) GIS & Middle Earth Presentation & Data Set. William & Mary. https://doi.org/10.21220/RKEZ-X707
}
\usage{
lotr_lyr
}
\description{
RasterLayer of middle earth based on an example digital elevation model of Tolkien’s
Middle Earth produced by the Center for Geospatial Analysis at William & Mary
(\href{https://scholarworks.wm.edu/asoer/3/}{Robert, 2020}).
}
\keyword{datasets}
