% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lotr_range}
\alias{lotr_range}
\title{Middle earth example range polygon}
\format{
sf
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
lotr_range
}
\description{
sf polygon of range map
}
\keyword{datasets}
