% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_gd.R
\name{mask_gd}
\alias{mask_gd}
\title{Mask diversity map based on sample number}
\usage{
mask_gd(
  x,
  min_n,
  plot = FALSE,
  bkg.col = "white",
  col.pal = viridis::magma(100)
)
}
\arguments{
\item{x}{RasterStack where the first layer is genetic diversity and the second layer is sample count}

\item{min_n}{minimum number of samples (everything less than this number is masked)}

\item{plot}{whether to plot results}

\item{bkg.col}{background color for plotting map}

\item{col.pal}{color palette to use for plotting}
}
\value{
RasterLayer
}
\description{
Mask diversity map based on sample number
}
\examples{

load_mini_ex()
wpi <- window_gd(mini_vcf, mini_coords, mini_lyr, nloci = 10, rarify = TRUE)
kpi <- krig_gd(wpi, mini_lyr)
mpi <- mask_gd(kpi, min_n = 3)
plot_gd(mpi, "Kriged and Masked Pi")

}
