% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_lyr}
\alias{mini_lyr}
\title{Mini middle earth example raster}
\format{
A SpatRaster of middle earth
}
\source{
created from simulations in Bishop et al., 202X
}
\usage{
mini_lyr
}
\description{
SpatRaster of middle earth from CITE
}
\keyword{datasets}
