% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_lyr}
\alias{mini_lyr}
\title{Mini middle earth example raster}
\format{
A RasterLayer of middle earth
}
\usage{
mini_lyr
}
\description{
RasterLayer of middle earth from CITE
}
\keyword{datasets}
