% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_lyr}
\alias{mini_lyr}
\title{Mini middle earth example raster}
\format{
A RasterLayer of middle earth
}
\source{
created from simulations in Bishop et al. (2023)
}
\usage{
mini_lyr
}
\description{
Small RasterLayer of middle earth based on an example digital elevation model of Tolkien's
Middle Earth produced by the Center for Geospatial Analysis at William & Mary
(\href{https://scholarworks.wm.edu/asoer/3/}{Robert, 2020}).
}
\keyword{datasets}
