% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gd.R
\name{plot_count}
\alias{plot_count}
\title{Plot sample counts}
\usage{
plot_count(
  x,
  index = NULL,
  breaks = 10,
  col = viridis::mako(breaks),
  zlim = NULL,
  main = NULL,
  legend = TRUE,
  legend.width = 1,
  axis.args = list(cex.axis = 1)
)
}
\arguments{
\item{x}{RasterLayer of counts or RasterStack where indexed layer is sample counts}

\item{index}{if RasterStack is provided, index of the sample count layer to plot (assumes this is a stacked output from window_gd and defaults to plotting second layer)}

\item{breaks}{number of breaks to use in color scale (defaults to 10)}

\item{col}{color pallete to use for plotting (defaults to viridis::magma pallete)}

\item{zlim}{limits of the color scale values}

\item{main}{character. Main plot title}

\item{legend}{if FALSE, no legend bar is displayed.}

\item{legend.width}{Width in characters of the legend bar}

\item{axis.args}{A list of additional arguments for the legend axis}
}
\value{
plot of sample counts
}
\description{
Plot sample counts
}
\examples{
data("mini_lyr")
plot_count(mini_lyr)
}
