% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_count}
\alias{plot_count}
\title{Plot sample counts}
\usage{
plot_count(
  x,
  col = viridis::mako(100),
  zlim = NULL,
  main = NULL,
  legend = legend
)
}
\arguments{
\item{x}{output from \link[wingen]{window_gd} or \link[wingen]{krig_gd} (RasterStack where second layer is sample counts)}

\item{col}{A color palette, i.e. a vector of n contiguous colors generated by functions like \link{rainbow}, \link{heat.colors}, \link{topo.colors}, \link[sp]{bpy.colors} or one or your own making, perhaps using \code{\link{colorRampPalette}}. If none is provided, \code{rev(terrain.colors(255))} is used unless \code{x} has a 'color table'}

\item{zlim}{limits of the color scale values}

\item{main}{character. Main plot title}
}
\value{
plot of sample counts
}
\description{
Plot sample counts
}
