% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gd.R
\name{plot_count}
\alias{plot_count}
\title{Plot sample counts}
\usage{
plot_count(
  x,
  index = 2,
  breaks = 10,
  col = viridis::mako(breaks),
  zlim = NULL,
  main = NULL,
  legend = TRUE,
  legend.width = 1,
  axis.args = list(cex.axis = 1)
)
}
\arguments{
\item{x}{RasterLayer of counts or RasterStack where indexed layer is sample counts}

\item{index}{if RasterStack is provided, index of the sample count layer to plot (defaults to 2)}

\item{breaks}{number of breaks to use in color scale (defaults to 10)}

\item{col}{color pallete to use for plotting (defaults to viridis::magma pallete)}

\item{zlim}{limits of the color scale values}

\item{main}{character. Main plot title}
}
\value{
plot of sample counts
}
\description{
Plot sample counts
}
