% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gd}
\alias{plot_gd}
\title{Plot genetic diversity results}
\usage{
plot_gd(x, col = viridis::magma(100), zlim = NULL, main = NULL)
}
\arguments{
\item{x}{output from \link[wingen]{window_gd} or \link[wingen]{krig_gd} (RasterStack where first layer is genetic diversity)}

\item{col}{A color palette, i.e. a vector of n contiguous colors generated by functions like \link{rainbow}, \link{heat.colors}, \link{topo.colors}, \link[sp]{bpy.colors} or one or your own making, perhaps using \code{\link{colorRampPalette}}. If none is provided, \code{rev(terrain.colors(255))} is used unless \code{x} has a 'color table'}

\item{main}{character. Main plot title}
}
\value{
plot of genetic diversity
}
\description{
Plot genetic diversity results
}
