% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_gd.R
\name{preview_gd}
\alias{preview_gd}
\title{Preview moving window and sample counts}
\usage{
preview_gd(
  lyr,
  coords,
  wdim,
  fact = 0,
  sample_count = TRUE,
  min_n = 0,
  plot = TRUE
)
}
\arguments{
\item{lyr}{SpatRaster or RasterLayer to slide the window across (see Details for important information about projections)}

\item{coords}{coordinates of samples as sf points, a two-column matrix, or a data.frame representing x and y coordinates (see Details for important information about projections)}

\item{wdim}{dimensions (height x width) of window; if only one value is provided, a square window is created (defaults to 3 x 3 window)}

\item{fact}{aggregation factor to apply to \code{lyr} (defaults to 0; \emph{note:} increasing this value reduces computational time)}

\item{sample_count}{whether to create plot of sample counts for each cell (defaults to TRUE)}

\item{min_n}{min number of samples to use in calculations (any focal cell with a window containing less than this number of samples will be assigned a value of NA)}

\item{plot}{whether to plot results (default = TRUE)}
}
\value{
SpatRaster with sample counts layer (if sample_count = TRUE)
}
\description{
Generate preview of moving window size and sample counts based on the coordinates and parameters to be supplied to \link[wingen]{window_gd}
}
\details{
Coordinates and rasters should be in a projected (planar) coordinate system such that raster cells are of equal sizes.
Therefore, spherical systems (including latitute-longitude coordinate systems) should be projected prior to use.
Transformation can be performed using \link[sf]{st_set_crs} for coordinates or \link[terra]{project} for rasters (see vignette for more details).
}
\examples{
load_mini_ex()
preview_gd(mini_lyr, mini_coords, wdim = 3, fact = 3, sample_count = TRUE, min_n = 2)
}
