% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{rarify_gd}
\alias{rarify_gd}
\title{Helper function to rarify subsample and calculate genetic diversity}
\usage{
rarify_gd(gen, sub, rarify_nit = 10, rarify_n = 4, stat, fun, L = NULL)
}
\arguments{
\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{stat}{function to calculate genetic diversity (can either be calc_mean_ar, calc_pi, calc_mean_biar, or calc_mean_het)}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}

\item{L}{for calculating pi, L argument in \link[hierfstat]{pi.dosage} function. Return the average nucleotide diversity per nucleotide given the length L of the sequence. The wingen defaults is L = "nvariants" which sets L to the number of variants in the VCF. If L = NULL, returns the sum over SNPs of nucleotide diversity (note: L = NULL is the \link[hierfstat]{pi.dosage} default which wingen does not to use).}
}
\value{
rarified genetic diversity statistic
}
\description{
Helper function to rarify subsample and calculate genetic diversity
}
\keyword{internal}
