% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{rarify_helper}
\alias{rarify_helper}
\title{Rarefaction helper function}
\usage{
rarify_helper(gen, sub, rarify_n, rarify_nit, stat, fun = mean, nloci = NULL)
}
\arguments{
\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}

\item{nloci}{for calculating pi, if nloci=NULL (default), returns the sum over SNPs of nucleotide diversity; otherwise return the average nucleotide diversity per nucleotide given the length nloci of the sequence (L argument in \link[hierfstat]{pi.dosage} function)}
}
\description{
Rarefaction helper function
}
\keyword{internal}
