% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{raster_to_grid}
\alias{raster_to_grid}
\title{Conver a raster to a grid}
\usage{
raster_to_grid(x, agg = NULL, disagg = NULL)
}
\arguments{
\item{x}{RasterLayer}

\item{agg}{factor used for aggregation if provided}

\item{disagg}{factor used for disaggregation if provided}
}
\value{
gridded SpatialPixelsDataFrame
}
\description{
Conver a raster to a grid
}
