% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_div.R
\name{raster_to_grid}
\alias{raster_to_grid}
\title{Conver a raster to a grid}
\usage{
raster_to_grid(x, agg = NULL, disagg = NULL)
}
\arguments{
\item{x}{RasterLayer}

\item{fact}{factor for aggregation (defaults to no aggregation)}
}
\value{
gridded SpatialPixelsDataFrame
}
\description{
Conver a raster to a grid
}
