% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{sample_gd}
\alias{sample_gd}
\title{Helper function to calculate genetic diversity of a sample}
\usage{
sample_gd(gen, sub, stat, L = NULL)
}
\arguments{
\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{stat}{function to calculate genetic diversity (can either be calc_mean_arcalc_pi, calc_mean_biar, or calc_mean_het)}

\item{L}{for calculating pi, L argument in \link[hierfstat]{pi.dosage} function. Return the average nucleotide diversity per nucleotide given the length L of the sequence. The wingen defaults is L = "nvariants" which sets L to the number of variants in the VCF. If L = NULL, returns the sum over SNPs of nucleotide diversity (note: L = NULL is the \link[hierfstat]{pi.dosage} default which wingen does not to use).}
}
\value{
mean allelic richness of a subsample
}
\description{
Helper function to calculate genetic diversity of a sample
}
\keyword{internal}
