% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{sample_gd}
\alias{sample_gd}
\title{Helper function to calculate genetic diversity of a sample}
\usage{
sample_gd(gen, sub, stat, nloci = NULL)
}
\arguments{
\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{nloci}{if nloci=NULL (default), returns the sum over SNPs of nucleotide diversity; otherwise return the average nucleotide diversity per nucleotide given the length nloci of the sequence (L argument in \link[hierfstat]{pi.dosage} function)}
}
\value{
mean allelic richness of a subsample
}
\description{
Helper function to calculate genetic diversity of a sample
}
\keyword{internal}
