% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{window_gd}
\alias{window_gd}
\title{Sliding window map of genetic diversity}
\usage{
window_gd(
  vcf,
  coords,
  lyr,
  stat = "pi",
  fact = 0,
  wdim = 5,
  rarify = FALSE,
  rarify_n = 4,
  rarify_nit = 5,
  min_n = 2,
  fun = mean,
  parallel = FALSE,
  nloci = NULL
)
}
\arguments{
\item{vcf}{object of type vcf ( (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.))}

\item{coords}{coordinates (two columns, the first should be x and the second should be y and the order should be the same as the genetic data),}

\item{lyr}{RasterLayer to slide window across}

\item{stat}{genetic diversity stat to calculate (can either be "pi" for nucleotide diversity, "het" for average heterozygosity across all loci, "allelic.richness" for average allelic richness across all loci, or "biallelic.richness" to get average allelic richness across all loci for a biallelic dataset (this option faster than "allelic.richness"))}

\item{fact}{aggregation factor to apply to the RasterLayer (\emph{note:} increasing this value reduces computational time)}

\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created}

\item{rarify}{if rarify = TRUE, rarefaction is performed}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{min_n}{min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}

\item{parallel}{whether to parallelize the function (see vignette for setting up a cluster to do so)}

\item{nloci}{for calculating pi, if nloci=NULL (default), returns the sum over SNPs of nucleotide diversity; otherwise return the average nucleotide diversity per nucleotide given the length nloci of the sequence (L argument in \link[hierfstat]{pi.dosage} function)}
}
\value{
RasterStack that includes a raster of genetic diversity and a raster of the number of samples within the window for each cell
}
\description{
Sliding window map of genetic diversity
}
\examples{
load_mini_ex()
wpi <- window_gd(vcf, coords, lyr, stat = "pi", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
whet <- window_gd(vcf, coords, lyr, stat = "het", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
war <- window_gd(vcf, coords, lyr, stat = "biallelic.richness", nloci = 10, rarify_n = 4, rarify_nit = 5, rarify = TRUE)
plot_gd(wpi)
plot_gd(whet)
plot_gd(war)
plot_count(wpi)

}
