% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{window_gd_general}
\alias{window_gd_general}
\title{Helper function for window_gd}
\usage{
window_gd_general(
  gen,
  coords,
  lyr,
  stat = calc_mean_ar,
  fact = 0,
  wdim = 10,
  rarify = FALSE,
  rarify_n = 4,
  rarify_nit = 10,
  min_n = 2,
  fun = mean,
  parallel = FALSE
)
}
\arguments{
\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{coords}{coordinates (two columns, the first should be x and the second should be y and the order should be the same as the genetic data),}

\item{lyr}{RasterLayer to slide window across}

\item{fact}{aggregation factor to apply to the RasterLayer (\emph{note:} increasing this value reduces computational time)}

\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created}

\item{rarify}{if rarify = TRUE, rarefaction is performed}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{min_n}{min number of samples to calculate allelic richness for (equal to rarify_n if provided, otherwise defaults to 2)}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}
}
\value{
RasterStack that includes a raster of genetic diversity and a raster of the number of samples within the window for each cell
}
\description{
Helper function for window_gd
}
\keyword{internal}
