% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{window_general}
\alias{window_general}
\title{General function for making moving window maps}
\usage{
window_general(
  x,
  coords,
  lyr,
  stat,
  wdim = 3,
  fact = 0,
  rarify = FALSE,
  rarify_n = 2,
  rarify_nit = 10,
  min_n = 2,
  fun = mean,
  L = "nvariants",
  rarify_alleles = TRUE,
  parallel = FALSE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{x}{data to be summarized by the moving window (\emph{note:} order matters! \code{coords} should be in the same order, there are currently no checks for this). The class of \code{x} required depends on the statistic being calculated (see the \code{stat} argument and the function description for more details)}

\item{coords}{two-column matrix or data.frame representing x (longitude) and y (latitude) coordinates of samples}

\item{lyr}{RasterLayer to slide the window across}

\item{stat}{moving window statistic to calculate (can either be \code{pi} for nucleotide diversity (\code{x} must be a dosage matrix), \code{Ho} for average observed heterozygosity across all loci (\code{x} must be a heterozygosity matrix) , "allelic_richness" for average allelic richness across all loci (\code{x} must be a \code{genind} type object), "biallelic_richness" to get average allelic richness across all loci for a biallelic dataset (\code{x} must be a dosage matrix). \code{stat} can also be set to any function that will take \code{x}as input and return a single numeric value (for example, \code{x} can be a vector and \code{stat} can be set equal to a summary statistic like \code{mean}, \code{sum}, or \code{sd})}

\item{wdim}{dimensions (height x width) of window, if only one value is provided a square window is created (defaults to 3 x 3 window)}

\item{fact}{aggregation factor to apply to the RasterLayer (defaults to 0; \emph{note:} increasing this value reduces computational time)}

\item{rarify}{if rarify = TRUE, rarefaction is performed (defaults to FALSE)}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction (defaults to 2)}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction (defaults to 5)}

\item{min_n}{min number of samples to use in calculations (any focal cell with a window containing less than this number of samples will be assigned a value of NA; equal to rarify_n if rarify = TRUE, otherwise defaults to 2)}

\item{fun}{function to use to summarize rarefaction results (defaults to mean)}

\item{L}{for calculating pi, L argument in \link[hierfstat]{pi.dosage} function. Return the average nucleotide diversity per nucleotide given the length L of the sequence. The wingen defaults is L = "nvariants" which sets L to the number of variants in the VCF. If L = NULL, returns the sum over SNPs of nucleotide diversity (note: L = NULL is the \link[hierfstat]{pi.dosage} default which wingen does not to use).}

\item{rarify_alleles}{for calculating biallelic_richness, whether to perform rarefaction of allele counts as in \link[hierfstat]{allelic.richness} (defaults to TRUE)}

\item{parallel}{whether to parallelize the function (defaults to FALSE)}

\item{ncores}{if parallel = TRUE, number of cores to use for parallelization (defaults to total available number of cores minus 1)}

\item{...}{if a function is provided for \code{stat}, additional arguments to pass to the \code{stat} function (e.g. if \code{stat = mean}, users may want to set \code{na.rm = TRUE})}
}
\value{
RasterStack that includes a raster of genetic diversity and a raster of the number of samples within the window for each cell
}
\description{
Generate a continuous raster map using moving windows. While \link[wingen]{window_gd} is built specifically for making moving window maps of genetic diversity from vcfs,
\code{window_general} can be used to make moving window maps from different data inputs. Unlike \code{window_gd}, \code{window_general} will not convert your data into
the correct format for calculations of different diversity metrics. To calculate \code{pi} or \code{biallelic_richness}, \code{x} must be a dosage matrix with values of 0, 1, or 2 To calculate
\code{Ho}, \code{x} must be a heterozygosity matrix where values of 0 = homozygosity and values of 1 = heterozygosity. To calculate \code{allelic_richness}, \code{x} must be a \code{genind} type object.
Users can set \code{x} to a vector and create moving window maps with any function that can be applied to a vector (e.g. \code{stat = mean}, \code{var}, \code{sum}, etc.).
\code{x} can also be a matrix or data frame (where rows are individuals), and then \code{stat} can be any function that takes a matrix or data frame and outputs a
single numeric value (e.g., a function that produces a custom diversity index), however this should be attempted with caution since this functionality has
not have been tested extensively and may produce errors.
}
