% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gd.R
\name{window_helper}
\alias{window_helper}
\title{Helper function for window calculations}
\usage{
window_helper(
  i,
  lyr,
  gen,
  coord_cells,
  nmat,
  stat,
  rarify,
  rarify_n,
  rarify_nit,
  min_n,
  fun,
  L = NULL
)
}
\arguments{
\item{i}{cell index}

\item{lyr}{RasterLayer to slide window across}

\item{gen}{genetic data (\emph{note:} order matters! the coordinate and genetic data should be in the same order, there are currently no checks for this.)}

\item{coord_cells}{cell indices for each coordinate}

\item{nmat}{neighborhood matrix}

\item{stat}{function to calculate genetic diversity (can either be calc_mean_ar, calc_pi, calc_mean_biar, or calc_mean_het)}

\item{rarify}{if rarify = TRUE, rarefaction is performed}

\item{rarify_n}{if rarify = TRUE, number of points to use for rarefaction}

\item{rarify_nit}{if rarify = TRUE, number of iterations to use for rarefaction}

\item{min_n}{min number of samples to use in calculations (any focal cell with a window containing less than this number of samples will be assigned a value of NA; equal to rarify_n if rarify = TRUE, otherwise defaults to 2)}

\item{fun}{function to use to summarize data in window (defaults to base R mean)}

\item{L}{for calculating pi, L argument in \link[hierfstat]{pi.dosage} function. Return the average nucleotide diversity per nucleotide given the length L of the sequence. The wingen defaults is L = "nvariants" which sets L to the number of variants in the VCF. If L = NULL, returns the sum over SNPs of nucleotide diversity (note: L = NULL is the \link[hierfstat]{pi.dosage} default which wingen does not to use).}
}
\value{
genetic diversity and counts for a single cell
}
\description{
Helper function for window calculations
}
\keyword{internal}
