% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingen-package.R
\docType{package}
\name{wingen-package}
\alias{wingen}
\alias{wingen-package}
\title{wingen: Continuous Mapping of Genetic Diversity}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate continuous maps of genetic diversity using moving windows with options for rarefaction, interpolation, and masking as described in Bishop et al. (2023) \doi{10.1111/2041-210X.14090}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnushaPB/wingen}
  \item Report bugs at \url{https://github.com/AnushaPB/wingen/issues}
}

}
\author{
\strong{Maintainer}: Anusha Bishop \email{anusha.bishop@berkeley.edu} (\href{https://orcid.org/0000-0003-1731-8683}{ORCID})

Authors:
\itemize{
  \item Anne Chambers \email{eachambers@berkeley.edu} (\href{https://orcid.org/0000-0002-7369-0108}{ORCID})
  \item Ian Wang \email{ianwang@berkeley.edu} (\href{https://orcid.org/0000-0003-2554-9414}{ORCID})
}

}
\keyword{internal}
